Move custom_detectors validation code into its own file
